On M1 Macs, long double is just double anyway, so don't use it

+++ geometry.cpp
@@ -160,25 +160,25 @@ drawvec remove_noop(drawvec geom, int type, int shift) {
 	return out;
 }
 
-long double get_area(drawvec &geom, size_t i, size_t j) {
+double get_area(drawvec &geom, size_t i, size_t j) {
 	// Coordinates in `geom` are 40-bit integers, so there is no good way
 	// to multiply them without possible precision loss. Since they probably
 	// do not use the full precision, shift them nearer to the origin so
-	// their product is more likely to be exactly representable as a long double.
+	// their product is more likely to be exactly representable as a double.
 	long long bx = geom[i].x;
 	long long by = geom[i].y;
 
-	long double area = 0;
+	double area = 0;
 	for (size_t k = i; k < j; k++) {
-		area += ((long double) (geom[k].x - bx) * (long double) (geom[i + ((k - i + 1) % (j - i))].y - by)) -
-		        ((long double) (geom[k].y - by) * (long double) (geom[i + ((k - i + 1) % (j - i))].x - bx));
+		area += ((double) (geom[k].x - bx) * (double) (geom[i + ((k - i + 1) % (j - i))].y - by)) -
+		        ((double) (geom[k].y - by) * (double) (geom[i + ((k - i + 1) % (j - i))].x - bx));
 	}
 	area /= 2;
 	return area;
 }
 
-long double get_mp_area(drawvec &geom) {
-	long double ret = 0;
+double get_mp_area(drawvec &geom) {
+	double ret = 0;
 
 	for (size_t i = 0; i < geom.size(); i++) {
 		if (geom[i].op == VT_MOVETO) {
@@ -214,10 +214,10 @@ static void decode_clipped(mapbox::geometry::multi_polygon<long long> &t, drawve
 				ring.push_back(draw(VT_LINETO, ring[0].x, ring[0].y));
 			}
 
-			long double area = get_area(ring, 0, ring.size());
+			double area = get_area(ring, 0, ring.size());
 
 			if ((j == 0 && area < 0) || (j != 0 && area > 0)) {
-				fprintf(stderr, "Ring area has wrong sign: %Lf for %zu\n", area, j);
+				fprintf(stderr, "Ring area has wrong sign: %f for %zu\n", area, j);
 				exit(EXIT_IMPOSSIBLE);
 			}
 
@@ -402,7 +402,7 @@ void check_polygon(drawvec &geom) {
 				}
 			}
 
-			long double area = get_area(geom, i, j);
+			double area = get_area(geom, i, j);
 
 #if 0
 			fprintf(stderr, "looking at %lld to %lld, area %f\n", (long long) i, (long long) j, area);
@@ -523,9 +523,9 @@ drawvec simple_clip_poly(drawvec &geom, int z, int buffer) {
 	return simple_clip_poly(geom, -clip_buffer, -clip_buffer, area + clip_buffer, area + clip_buffer);
 }
 
-drawvec reduce_tiny_poly(drawvec &geom, int z, int detail, bool *reduced, long double *accum_area) {
+drawvec reduce_tiny_poly(drawvec &geom, int z, int detail, bool *reduced, double *accum_area) {
 	drawvec out;
-	const long double pixel = (1LL << (32 - detail - z)) * (long double) tiny_polygon_size;
+	const double pixel = (1LL << (32 - detail - z)) * (double) tiny_polygon_size;
 
 	*reduced = true;
 	bool included_last_outer = false;
@@ -539,7 +539,7 @@ drawvec reduce_tiny_poly(drawvec &geom, int z, int detail, bool *reduced, long d
 				}
 			}
 
-			long double area = get_area(geom, i, j);
+			double area = get_area(geom, i, j);
 
 			// XXX There is an ambiguity here: If the area of a ring is 0 and it is followed by holes,
 			// we don't know whether the area-0 ring was a hole too or whether it was the outer ring
@@ -964,7 +964,7 @@ drawvec fix_polygon(drawvec &geom) {
 				// GeoJSON winding is reversed from vector winding
 				reverse_ring = false;
 			} else {
-				long double area = get_area(ring, 0, ring.size());
+				double area = get_area(ring, 0, ring.size());
 				if ((area > 0) != outer) {
 					reverse_ring = true;
 				}
